Move GeffReader into core_io._base_read
