+++ src/geff/core_io/__init__.py
@@ -1,0 +1,3 @@
+from ._base_read import GeffReader, read_to_memory
+
+__all__ = ["GeffReader", "read_to_memory"]
