+++ src/geff/graph_libs/_networkx.py
@@ -4,7 +4,7 @@ from typing import TYPE_CHECKING, Any, Literal
 
 import networkx as nx
 
-from geff.geff_reader import read_to_memory
+from geff.core_io._base_read import read_to_memory
 from geff.io_utils import (
     calculate_roi_from_nodes,
     create_or_update_metadata,
